Extract test utility createLevelStateLazy
